ThreatKit/requirements.txt
Flask>=3.0
zxcvbn
transformers>=4.44
torch>=2.0
joblib
scikit-learn
pefile
pandas
numpy
yara-python
requests
dnspython